Improve: add FollowingHandler to handle all replication requests
