+++ openraft/src/engine/engine_impl.rs
@@ -1,6 +1,5 @@
-use std::sync::Arc;
-
 use crate::core::ServerState;
+use crate::engine::handler::following_handler::FollowingHandler;
 use crate::engine::handler::log_handler::LogHandler;
 use crate::engine::handler::replication_handler::ReplicationHandler;
 use crate::engine::handler::server_state_handler::ServerStateHandler;
@@ -22,7 +21,6 @@ use crate::raft::VoteRequest;
 use crate::raft::VoteResponse;
 use crate::raft_state::LogStateReader;
 use crate::raft_state::RaftState;
-use crate::raft_types::RaftLogId;
 use crate::summary::MessageSummary;
 use crate::validate::Valid;
 use crate::LogId;
@@ -410,7 +408,7 @@ where
         self.output.push_command(Command::MoveInputCursorBy { n: l });
     }
 
-    // TODO: AcceptorHandler
+    // TODO: move logic to FollowingHandler
     /// Append entries to follower/learner.
     ///
     /// Also clean conflicting entries and update membership state.
@@ -451,7 +449,7 @@ where
                 let local = self.state.get_log_id(prev.index);
                 tracing::debug!(local = debug(&local), "prev_log_id does not match");
 
-                self.truncate_logs(prev.index);
+                self.following_handler().truncate_logs(prev.index);
                 return AppendEntriesResponse::Conflict;
             }
         }
@@ -464,135 +462,19 @@ where
         );
 
         let l = entries.len();
-        let since = self.first_conflicting_index(entries);
+        let since = self.following_handler().first_conflicting_index(entries);
         if since < l {
             // Before appending, if an entry overrides an conflicting one,
             // the entries after it has to be deleted first.
             // Raft requires log ids are in total order by (term,index).
             // Otherwise the log id with max index makes committed entry invisible in election.
-            self.truncate_logs(entries[since].get_log_id().index);
-            self.follower_do_append_entries(entries, since);
+            self.following_handler().truncate_logs(entries[since].get_log_id().index);
+            self.following_handler().follower_do_append_entries(entries, since);
         }
 
-        self.follower_commit_entries(leader_committed, prev_log_id, entries);
+        self.following_handler().follower_commit_entries(leader_committed, prev_log_id, entries);
 
         AppendEntriesResponse::Success
-    }
-
-    // TODO: AcceptorHandler
-    #[tracing::instrument(level = "debug", skip_all)]
-    pub(crate) fn follower_commit_entries<'a, Ent: RaftEntry<NID, N> + 'a>(
-        &mut self,
-        leader_committed: Option<LogId<NID>>,
-        prev_log_id: Option<LogId<NID>>,
-        entries: &[Ent],
-    ) {
-        tracing::debug!(
-            leader_committed = display(leader_committed.summary()),
-            prev_log_id = display(prev_log_id.summary()),
-        );
-
-        // Committed index can not > last_log_id.index
-        let last = entries.last().map(|x| *x.get_log_id());
-        let last = std::cmp::max(last, prev_log_id);
-        let committed = std::cmp::min(leader_committed, last);
-
-        tracing::debug!(committed = display(committed.summary()), "update committed");
-
-        if let Some(prev_committed) = self.state.update_committed(&committed) {
-            self.output.push_command(Command::FollowerCommit {
-                // TODO(xp): when restart, commit is reset to None. Use last_applied instead.
-                already_committed: prev_committed,
-                upto: committed.unwrap(),
-            });
-        }
-
-        // TODO(5): follower has not yet commit the membership_state.
-        //          For now it is OK. But it should be done here.
-    }
-
-    // TODO: AcceptorHandler
-    /// Follower/Learner appends `entries[since..]`.
-    ///
-    /// It assumes:
-    /// - Previous entries all match.
-    /// - conflicting entries are deleted.
-    ///
-    /// Membership config changes are also detected and applied here.
-    #[tracing::instrument(level = "debug", skip(self, entries))]
-    pub(crate) fn follower_do_append_entries<'a, Ent: RaftEntry<NID, N> + 'a>(
-        &mut self,
-        entries: &[Ent],
-        since: usize,
-    ) {
-        let l = entries.len();
-        if since == l {
-            return;
-        }
-
-        let entries = &entries[since..];
-
-        debug_assert_eq!(
-            entries[0].get_log_id().index,
-            self.state.log_ids.last().cloned().next_index(),
-        );
-
-        debug_assert!(Some(entries[0].get_log_id()) > self.state.log_ids.last());
-
-        self.state.extend_log_ids(entries);
-
-        self.output.push_command(Command::AppendInputEntries { range: since..l });
-
-        self.follower_append_membership(entries.iter());
-
-        // TODO(xp): should be moved to handle_append_entries_req()
-        self.output.push_command(Command::MoveInputCursorBy { n: l });
-    }
-
-    // TODO: AcceptorHandler
-    /// Delete log entries since log index `since`, inclusive, when the log at `since` is found conflict with the
-    /// leader.
-    ///
-    /// And revert effective membership to the last committed if it is from the conflicting logs.
-    #[tracing::instrument(level = "debug", skip(self))]
-    pub(crate) fn truncate_logs(&mut self, since: u64) {
-        tracing::debug!(since = since, "truncate_logs");
-
-        debug_assert!(since >= self.state.last_purged_log_id().next_index());
-
-        let since_log_id = match self.state.get_log_id(since) {
-            None => {
-                tracing::debug!("trying to delete absent log at: {}", since);
-                return;
-            }
-            Some(x) => x,
-        };
-
-        self.state.log_ids.truncate(since);
-        self.output.push_command(Command::DeleteConflictLog { since: since_log_id });
-
-        let changed = self.state.membership_state.truncate(since);
-        if let Some(c) = changed {
-            self.output.push_command(Command::UpdateMembership { membership: c });
-            self.server_state_handler().update_server_state_if_changed();
-        }
-    }
-
-    // TODO: AcceptorHandler
-    /// Update membership state with a committed membership config
-    #[tracing::instrument(level = "debug", skip_all)]
-    pub(crate) fn update_committed_membership(&mut self, membership: EffectiveMembership<NID, N>) {
-        tracing::debug!("update committed membership: {}", membership.summary());
-
-        let m = Arc::new(membership);
-
-        // TODO: if effective membership changes, call `update_repliation()`
-        let effective_changed = self.state.membership_state.update_committed(m);
-        if let Some(c) = effective_changed {
-            self.output.push_command(Command::UpdateMembership { membership: c })
-        }
-
-        self.server_state_handler().update_server_state_if_changed();
     }
 
     /// Leader steps down(convert to learner) once the membership not containing it is committed.
@@ -620,110 +502,6 @@ where
                 self.vote_handler().become_following();
             }
         }
-    }
-
-    // TODO: AcceptorHandler
-    /// Follower/Learner handles install-snapshot.
-    #[tracing::instrument(level = "debug", skip_all)]
-    pub(crate) fn install_snapshot(&mut self, meta: SnapshotMeta<NID, N>) {
-        // There are two special cases in which snapshot last log id does not exists locally:
-        // Snapshot last log id before the local last-purged-log-id, or after the local last-log-id:
-        //
-        //      snapshot ----.
-        //                   v
-        // -----------------------llllllllll--->
-        //
-        //      snapshot ----.
-        //                   v
-        // ----lllllllllll--------------------->
-        //
-        // In the first case, snapshot-last-log-id <= last-purged-log-id <= local-snapshot-last-log-id.
-        // Thus snapshot is obsolete and won't be installed.
-        //
-        // In the second case, all local logs will be purged after install.
-
-        tracing::info!("install_snapshot: meta:{:?}", meta);
-
-        // TODO: temp solution: committed is updated after snapshot_last_log_id.
-        //       committed should be updated first or together with snapshot_last_log_id(i.e., extract `state` first).
-        let old_validate = self.state.enable_validate;
-        self.state.enable_validate = false;
-
-        let snap_last_log_id = meta.last_log_id;
-
-        if snap_last_log_id.as_ref() <= self.state.committed() {
-            tracing::info!(
-                "No need to install snapshot; snapshot last_log_id({}) <= committed({})",
-                snap_last_log_id.summary(),
-                self.state.committed().summary()
-            );
-            self.output.push_command(Command::CancelSnapshot { snapshot_meta: meta });
-            // TODO: temp solution: committed is updated after snapshot_last_log_id.
-            self.state.enable_validate = old_validate;
-            return;
-        }
-
-        // snapshot_last_log_id can not be None
-        let snap_last_log_id = snap_last_log_id.unwrap();
-
-        let mut snap_handler = self.snapshot_handler();
-        let updated = snap_handler.update_snapshot(meta.clone());
-        if !updated {
-            // TODO: temp solution: committed is updated after snapshot_last_log_id.
-            self.state.enable_validate = old_validate;
-            return;
-        }
-
-        // Do install:
-        // 1. Truncate all logs if conflict
-        //    Unlike normal append-entries RPC, if conflicting logs are found, it is not **necessary** to delete them.
-        //    But cleaning them make the assumption of incremental-log-id always hold, which makes it easier to debug.
-        //    See: [Snapshot-replication](https://datafuselabs.github.io/openraft/replication.html#snapshot-replication)
-        //
-        //    Truncate all:
-        //
-        //    It just truncate **ALL** logs here, because `snap_last_log_id` is committed, if the local log id conflicts
-        //    with `snap_last_log_id`, there must be a quorum that contains `snap_last_log_id`.
-        //    Thus it is safe to remove all logs on this node.
-        //
-        //    The logs before `snap_last_log_id` may conflicts with the leader too.
-        //    It's not safe to remove the conflicting logs that are less than `snap_last_log_id` after installing
-        //    snapshot.
-        //
-        //    If the node crashes, dirty logs may remain there. These logs may be forwarded to other nodes if this nodes
-        //    becomes a leader.
-        //
-        // 2. Install snapshot.
-
-        let local = self.state.get_log_id(snap_last_log_id.index);
-        if let Some(local) = local {
-            if local != snap_last_log_id {
-                // Delete non-committed logs.
-                self.truncate_logs(self.state.committed().next_index());
-            }
-        }
-
-        self.state.committed = Some(snap_last_log_id);
-        self.update_committed_membership(meta.last_membership.clone());
-
-        // TODO: There should be two separate commands for installing snapshot:
-        //       - Replace state machine with snapshot and replace the `current_snapshot` in the store.
-        //       - Do not install, just replace the `current_snapshot` with a newer one. This command can be used for
-        //         leader to synchronize its snapshot data.
-        self.output.push_command(Command::InstallSnapshot { snapshot_meta: meta });
-
-        // A local log that is <= snap_last_log_id can not conflict with the leader.
-        // But there will be a hole in the logs. Thus it's better remove all logs.
-
-        // In the second case, if local-last-log-id is smaller than snapshot-last-log-id,
-        // and this node crashes after installing snapshot and before purging logs,
-        // the log will be purged the next start up, in [`RaftState::get_initial_state`].
-        // TODO: move this to LogHandler::purge_log()?
-        self.state.purge_upto = Some(snap_last_log_id);
-        self.log_handler().purge_log();
-
-        // TODO: temp solution: committed is updated after snapshot_last_log_id.
-        self.state.enable_validate = old_validate;
     }
 
     #[tracing::instrument(level = "debug", skip_all)]
@@ -773,65 +551,6 @@ where
         rh.initiate_replication();
     }
 
-    // TODO: AcceptorHandler
-    /// Append membership log if membership config entries are found, after appending entries to log.
-    fn follower_append_membership<'a, Ent: RaftEntry<NID, N> + 'a>(
-        &mut self,
-        entries: impl DoubleEndedIterator<Item = &'a Ent>,
-    ) {
-        let memberships = Self::last_two_memberships(entries);
-        if memberships.is_empty() {
-            return;
-        }
-
-        // Update membership state with the last 2 membership configs found in new log entries.
-        // Other membership log can be just ignored.
-        for (i, m) in memberships.into_iter().enumerate() {
-            tracing::debug!(
-                last = display(m.summary()),
-                "applying {}-th new membership configs received from leader",
-                i
-            );
-            self.state.membership_state.append(Arc::new(m));
-        }
-
-        tracing::debug!(
-            membership_state = display(&self.state.membership_state.summary()),
-            "updated membership state"
-        );
-
-        self.output.push_command(Command::UpdateMembership {
-            membership: self.state.membership_state.effective().clone(),
-        });
-
-        self.server_state_handler().update_server_state_if_changed();
-    }
-
-    // TODO: AcceptorHandler
-    /// Find the last 2 membership entries in a list of entries.
-    ///
-    /// A follower/learner reverts the effective membership to the previous one,
-    /// when conflicting logs are found.
-    ///
-    /// See: [Effective-membership](https://datafuselabs.github.io/openraft/effective-membership.html)
-    fn last_two_memberships<'a, Ent: RaftEntry<NID, N> + 'a>(
-        entries: impl DoubleEndedIterator<Item = &'a Ent>,
-    ) -> Vec<EffectiveMembership<NID, N>> {
-        let mut memberships = vec![];
-
-        // Find the last 2 membership config entries: the committed and the effective.
-        for ent in entries.rev() {
-            if let Some(m) = ent.get_membership() {
-                memberships.insert(0, EffectiveMembership::new(Some(*ent.get_log_id()), m.clone()));
-                if memberships.len() == 2 {
-                    break;
-                }
-            }
-        }
-
-        memberships
-    }
-
     /// Check if a raft node is in a state that allows to initialize.
     ///
     /// It is allowed to initialize only when `last_log_id.is_none()` and `vote==(term=0, node_id=0)`.
@@ -860,31 +579,6 @@ where
         } else {
             Ok(())
         }
-    }
-
-    // TODO: AcceptorHandler
-    /// Find the first entry in the input that does not exist on local raft-log,
-    /// by comparing the log id.
-    fn first_conflicting_index<Ent: RaftLogId<NID>>(&self, entries: &[Ent]) -> usize {
-        let l = entries.len();
-
-        for (i, ent) in entries.iter().enumerate() {
-            let log_id = ent.get_log_id();
-            // for i in 0..l {
-            // let log_id = entries[i].get_log_id();
-
-            if !self.state.has_log_id(log_id) {
-                tracing::debug!(
-                    at = display(i),
-                    entry_log_id = display(log_id),
-                    "found nonexistent log id"
-                );
-                return i;
-            }
-        }
-
-        tracing::debug!("not found nonexistent");
-        l
     }
 
     fn assign_log_ids<'a, Ent: RaftEntry<NID, N> + 'a>(&mut self, entries: impl Iterator<Item = &'a mut Ent>) {
@@ -939,6 +633,16 @@ where
         ReplicationHandler {
             config: &mut self.config,
             leader,
+            state: &mut self.state,
+            output: &mut self.output,
+        }
+    }
+
+    pub(crate) fn following_handler(&mut self) -> FollowingHandler<NID, N> {
+        debug_assert!(self.internal_server_state.is_following());
+
+        FollowingHandler {
+            config: &mut self.config,
             state: &mut self.state,
             output: &mut self.output,
         }
