+++ openraft/src/engine/handler/mod.rs
@@ -1,3 +1,4 @@
+pub(crate) mod following_handler;
 pub(crate) mod log_handler;
 pub(crate) mod replication_handler;
 pub(crate) mod server_state_handler;
